Comment Controller Update

+++ apps/backend/src/comments/comments.service.ts
@@ -1,10 +1,19 @@
 import { Injectable } from '@nestjs/common';
 import { CommentsDBService, RestaurantDBService } from '@database';
-import { AddCommentsJsonDto, DeleteCommentsJsonDto } from './dtos';
+import {
+    AddCommentsJsonDto,
+    DeleteCommentsJsonDto,
+    UpdateCommentsJsonDto
+} from './dtos';
 import { Prisma as PrismaPostgres } from '@prisma/postgres/client';
 import { Prisma as PrismaMongoDb } from '@prisma/mongo/client';
 import { UserParamsDto } from './dtos/userparams.dto';
-import { BadRequestException, BadRequestExceptionType } from '@exceptions';
+import {
+    BadRequestException,
+    BadRequestExceptionType,
+    UnauthorizedException,
+    UnauthorizedExceptionType
+} from '@exceptions';
 import { ResponseController } from '@dtos';
 
 @Injectable()
@@ -106,9 +115,58 @@ export class CommentsService {
         data: DeleteCommentsJsonDto
     ): Promise<ResponseController> {
         const where: PrismaPostgres.CommentWhereUniqueInput = {
-            id: data.id
+            id: data.id,
+            user: { id: user.sub }
         };
+
+        const permission = await this.commentDBService.findUnique(where);
+
+        if (!permission) {
+            throw new UnauthorizedException(
+                UnauthorizedExceptionType.UNAUTHORIZED_ACCESS,
+                new Error('No permission to delete this comment'),
+                404
+            );
+        }
+
         const comment = await this.commentDBService.delete(where);
+        return {
+            Success: true,
+            Data: comment
+        };
+    }
+
+    async updateComment(
+        user: UserParamsDto,
+        updateData: UpdateCommentsJsonDto
+    ): Promise<ResponseController> {
+        const where: PrismaPostgres.CommentWhereUniqueInput = {
+            id: updateData.id,
+            user: {
+                id: user.sub
+            }
+        };
+
+        const permission = await this.commentDBService.findUnique(where);
+
+        if (!permission) {
+            throw new UnauthorizedException(
+                UnauthorizedExceptionType.UNAUTHORIZED_ACCESS,
+                new Error('No permission to delete this comment'),
+                404
+            );
+        }
+
+        const data: PrismaPostgres.CommentUpdateInput = {
+            ...updateData,
+            updatedAt: new Date()
+        };
+
+        const comment = await this.commentDBService.update({
+            where,
+            data
+        });
+
         return {
             Success: true,
             Data: comment
